Update vulkan to 1.1.73.0
Summary:
**loader: Fix a 1.0/1.1 versioning bug**
Fix a bug where when mixing a 1.0 and a 1.1 driver, if the loader stores the 1.0 driver first, the apiVersion field in vkCreateInstance will be rewritten to 1.0 for both drivers. Now it will only be changed for the 1.0 driver, as intended.

Full release notes available [here](https://vulkan.lunarg.com/doc/sdk/1.1.73.0/windows/release_notes.html)

Depends on D2791 and D2792

Signed-off-by: Pierre-Yves <pyu@riseup.net>

Test Plan: Unfortunately I cannot test vulkan in virtualbox.

Reviewers: #triage_team, ikey

Reviewed By: #triage_team, ikey

Subscribers: ikey

Differential Revision: https://dev.solus-project.com/D2793